Rename work card and list view items
