+++ assets/js/components/IngestSheet/Completed.jsx
@@ -9,8 +9,8 @@ import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 
 import Error from "../UI/Error";
 import IngestSheetCompletedErrors from "./Completed/Errors";
-import WorkListItem from "../Work/UIWorkListItem";
-import WorkCardItem from "../Work/UIWorkCardItem";
+import WorkListItem from "../Work/ListViewItem";
+import WorkCardItem from "../Work/CardViewItem";
 
 const IngestSheetCompleted = ({ sheetId }) => {
   const [isListView, setIsListView] = useState(false);
